+++ auth0/src/test/java/com/auth0/android/request/DefaultClientTest.kt
@@ -82,15 +82,30 @@ public class DefaultClientTest {
 
     @Test
     public fun shouldHaveLoggingDisabledByDefault() {
-        assertThat(DefaultClient().okHttpClient.interceptors, empty())
+        val netClient = DefaultClient(enableLogging = false)
+        assertThat(netClient.okHttpClient.interceptors, hasSize(1))
+        val interceptor: Interceptor = netClient.okHttpClient.interceptors[0]
+        assert(
+            interceptor is RetryInterceptor,
+        )
+    }
+
+    @Test
+    public fun shouldHaveRetryInterceptorEnabled() {
+        val netClient = DefaultClient(enableLogging = false)
+        assertThat(netClient.okHttpClient.interceptors, hasSize(1))
+        val interceptor: Interceptor = netClient.okHttpClient.interceptors[0]
+        assert(
+            interceptor is RetryInterceptor,
+        )
     }
 
     @Test
     public fun shouldHaveLoggingEnabledIfSpecified() {
         val netClient = DefaultClient(enableLogging = true)
-        assertThat(netClient.okHttpClient.interceptors, hasSize(1))
+        assertThat(netClient.okHttpClient.interceptors, hasSize(2))
 
-        val interceptor: Interceptor = netClient.okHttpClient.interceptors[0]
+        val interceptor: Interceptor = netClient.okHttpClient.interceptors[1]
         assertThat(
             (interceptor as HttpLoggingInterceptor).level,
             equalTo(HttpLoggingInterceptor.Level.BODY)